requirements.txt
Flask
Flask-SQLAlchemy
Flask-Mail
PyMySQL
python-dotenv
gunicorn
